requirement.txt
streamlit==1.39.0
pandas==2.2.2
numpy==1.26.4
scikit-learn==1.5.1
prophet==1.1.5
tensorflow==2.16.1
matplotlib==3.9.2
seaborn==0.13.2
fpdf2==2.8.1
openpyxl==3.1.5